Change place of calculation for timeincrement
Timeincrement will now be calculated in class EnergieSystem instead
of class BaseModel
